fix(inventory): fail-closed floors on physical ledger fields (#167–169)
Backstop behind the api-cloudrun server-side guards for the 2026-05-29 inventory
review (epic #175):
- inventory-ledger quantity_held/in_service/out_of_service + total_cost_basis +
  average_unit_cost: .min(0) (physical/valuation counts can't go negative).
- StoreBreakdown store + location quantity: .min(0).
- out-of-service create quantity: .int().positive().
- CreateProductInput.transaction: .refine() quantity == Σ transactionQuantity,
  quantity int().nonnegative().

Stock-summary quantity_available is deliberately left unconstrained — overbooking
is valid and shows negative availability.

+++ src/out-of-service.ts
@@ -257,7 +257,7 @@ export interface CreateOutOfServiceInputType {
 export const CreateOutOfServiceInput: z.ZodType<CreateOutOfServiceInputType> = z.object({
   uid_product: z.string().min(1),
   reason: OOSReasonEnum,
-  quantity: z.number().positive(),
+  quantity: z.number().int().positive(),
   dates: z.object({
     start: chicagoInstant().nullable().optional(),
     end: chicagoInstant().nullable().optional(),
